requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
fake-useragent>=1.4.0
python-docx>=0.8.11